Develop

+++ src/main/java/com/example/end/models/Cart.java
@@ -4,8 +4,6 @@ package com.example.end.models;
 import lombok.Data;
 import lombok.Getter;
 import lombok.Setter;
-import com.example.end.models.Procedure;
-
 
 import javax.persistence.*;
 import java.util.List;
@@ -18,7 +16,7 @@ public class Cart {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    private int id;
+    private Long id;
 
     @OneToOne
     @JoinColumn(name = "user_id")
@@ -43,4 +41,3 @@ public class Cart {
     }
 
 }
-
